requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
requests>=2.31.0
scipy>=1.11.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# AlphaGenome — install from source:
# git clone https://github.com/google-deepmind/alphagenome.git
# pip install ./alphagenome